refactor(app): extract WorkspacePage; App.tsx becomes thin router
Move all workspace UI logic from App.tsx into WorkspacePage.tsx, which
reads shared state from WorkspaceContext via useWorkspace(). App.tsx is
now a thin BrowserRouter/Routes shell. Add stub pages SavedConfigsPage
and OrderSummaryPage for future routes. Update App.test.tsx to render
WorkspacePage wrapped in WorkspaceProvider, preserving all 1085 tests.
