+++ packages/app/src/pages/OrderSummaryPage.tsx
@@ -1,0 +1,3 @@
+export function OrderSummaryPage() {
+  return <div>Order Summary (coming soon)</div>;
+}
