+++ packages/app/src/pages/SavedConfigsPage.tsx
@@ -1,0 +1,3 @@
+export function SavedConfigsPage() {
+  return <div>Saved Configs (coming soon)</div>;
+}
