+++ packages/app/src/pages/WorkspacePage.tsx
@@ -1,0 +1,491 @@
+import { useState, useEffect, useCallback, useRef, useMemo } from 'react';
+import type { ImageViewMode, DragData } from '../types/gridfinity';
+import { useWorkspace } from '../contexts/WorkspaceContext';
+import { useGridTransform } from '../hooks/useGridTransform';
+import { DimensionInput } from '../components/DimensionInput';
+import { GridPreview } from '../components/GridPreview';
+import { GridSummary } from '../components/GridSummary';
+import { ItemLibrary } from '../components/ItemLibrary';
+import { ItemControls } from '../components/ItemControls';
+import { BinCustomizationPanel } from '../components/BinCustomizationPanel';
+import { SpacerControls } from '../components/SpacerControls';
+import { BillOfMaterials } from '../components/BillOfMaterials';
+import { RefImageLibrary } from '../components/RefImageLibrary';
+import { ZoomControls } from '../components/ZoomControls';
+import { ImageViewToggle } from '../components/ImageViewToggle';
+import { GridViewport } from '../components/GridViewport';
+import { SidebarPanel } from '../components/SidebarPanel';
+import { LoadLayoutDialog } from '../components/layouts/LoadLayoutDialog';
+import { SubmissionsBadge } from '../components/admin/SubmissionsBadge';
+import { UserStlLibrarySection } from '../components/UserStlLibrarySection';
+import { exportToPdf } from '../utils/exportPdf';
+
+const API_BASE_URL = (import.meta.env.VITE_API_BASE_URL as string | undefined) ?? 'http://localhost:3001/api/v1';
+
+const LIBRARY_MIN_WIDTH = 160;
+const LIBRARY_MAX_WIDTH = 520;
+const LIBRARY_DEFAULT_WIDTH = 260;
+
+export function WorkspacePage() {
+  const ws = useWorkspace();
+
+  const {
+    width, setWidth, depth, setDepth, unitSystem, imperialFormat, setImperialFormat,
+    spacerConfig, setSpacerConfig, handleUnitChange,
+    gridResult, spacers,
+    placedItems, selectedItemIds,
+    rotateItem, deleteItem,
+    selectItem, selectAll, deselectAll, handleDrop, duplicateItem,
+    copyItems, pasteItems, deleteSelected, rotateSelected, updateItemCustomization,
+    bomItems,
+    layoutMeta, isReadOnly,
+    refImagePlacements, addRefImagePlacement, removeRefImagePlacement,
+    updateRefImagePosition, updateRefImageScale, updateRefImageOpacity,
+    updateRefImageRotation, toggleRefImageLock,
+    referenceImagesForGrid,
+    libraryItems, isLibraryLoading, isLibrariesLoading, libraryError, librariesError,
+    categories, getItemById, getLibraryMeta, refreshLibraries, refreshLibrary,
+    selectedLibraryMeta,
+    handleLoadLayout, handleSubmitClick, handleWithdrawLayout,
+    handleCloneCurrentLayout, handleClearAll, handleReset,
+    submitLayoutMutation, withdrawLayoutMutation, cloneLayoutMutation,
+    submittedCountQuery,
+    dialogs, dialogDispatch,
+    isAuthenticated, isAdmin,
+    exportPdfError, setExportPdfError,
+  } = ws;
+
+  // Local UI state
+  const [selectedImageId, setSelectedImageId] = useState<string | null>(null);
+  const [libraryTab, setLibraryTab] = useState<'items' | 'images'>('items');
+  const [libraryCategory, setLibraryCategory] = useState<string | null>(null);
+  const [libraryWidth, setLibraryWidth] = useState(LIBRARY_DEFAULT_WIDTH);
+  const libraryDragRef = useRef<{ startX: number; startWidth: number } | null>(null);
+
+  const [imageViewMode, setImageViewMode] = useState<ImageViewMode>(
+    () => (localStorage.getItem('gridfinity-image-view-mode') as ImageViewMode) || 'ortho'
+  );
+
+  const handleLibraryResizeStart = useCallback((e: React.MouseEvent) => {
+    e.preventDefault();
+    libraryDragRef.current = { startX: e.clientX, startWidth: libraryWidth };
+    const onMove = (ev: MouseEvent) => {
+      if (!libraryDragRef.current) return;
+      const delta = libraryDragRef.current.startX - ev.clientX;
+      setLibraryWidth(Math.min(LIBRARY_MAX_WIDTH, Math.max(LIBRARY_MIN_WIDTH, libraryDragRef.current.startWidth + delta)));
+    };
+    const onUp = () => {
+      libraryDragRef.current = null;
+      window.removeEventListener('mousemove', onMove);
+      window.removeEventListener('mouseup', onUp);
+    };
+    window.addEventListener('mousemove', onMove);
+    window.addEventListener('mouseup', onUp);
+  }, [libraryWidth]);
+
+  // Zoom and pan
+  const {
+    transform, zoomIn, zoomOut, resetZoom, fitToScreen,
+    handleWheel, setZoomLevel, pan,
+  } = useGridTransform();
+
+  const viewportRef = useRef<HTMLDivElement>(null);
+  const isSpaceHeldRef = useRef(false);
+
+  const handleFitToScreen = useCallback(() => {
+    const viewport = viewportRef.current;
+    if (!viewport) return;
+    const content = viewport.querySelector('.grid-preview') as HTMLElement | null;
+    if (!content) return;
+    const viewportRect = viewport.getBoundingClientRect();
+    fitToScreen(viewportRect.width, viewportRect.height, content.offsetWidth, content.offsetHeight);
+  }, [fitToScreen]);
+
+  const handleRotateSelectedCw = useCallback(() => { rotateSelected('cw'); }, [rotateSelected]);
+  const handleRotateSelectedCcw = useCallback(() => { rotateSelected('ccw'); }, [rotateSelected]);
+  const handleDeleteSelected = useCallback(() => { deleteSelected(); }, [deleteSelected]);
+
+  // Metadata map for broken state / image URLs
+  const refImageMetadata = useMemo(() => {
+    const map = new Map<string, { isBroken: boolean; imageUrl: string | null }>();
+    for (const p of refImagePlacements) {
+      map.set(p.id, {
+        isBroken: p.refImageId === null,
+        imageUrl: p.imageUrl ? `${API_BASE_URL}/images/${p.imageUrl}` : null,
+      });
+    }
+    return map;
+  }, [refImagePlacements]);
+
+  // Combined drop handler for both library items and ref images
+  const handleCombinedDrop = useCallback((dragData: DragData, x: number, y: number) => {
+    if (isReadOnly) return;
+    if (dragData.type === 'ref-image' && dragData.refImageId != null) {
+      const xPercent = (x / gridResult.gridX) * 100;
+      const yPercent = (y / gridResult.gridY) * 100;
+      addRefImagePlacement({
+        refImageId: dragData.refImageId, name: dragData.refImageName ?? 'Reference Image',
+        imageUrl: dragData.refImageUrl ?? '', x: xPercent, y: yPercent,
+        width: 25, height: 25, opacity: 0.5, scale: 1, isLocked: false, rotation: 0,
+      });
+    } else {
+      handleDrop(dragData, x, y);
+    }
+  }, [isReadOnly, gridResult.gridX, gridResult.gridY, addRefImagePlacement, handleDrop]);
+
+  const handleRefreshAll = useCallback(async () => {
+    try {
+      await refreshLibraries();
+      await refreshLibrary();
+    } catch (err) {
+      console.error('Library refresh failed:', err);
+    }
+  }, [refreshLibraries, refreshLibrary]);
+
+  const handleExportPdf = useCallback(async () => {
+    setExportPdfError(null);
+    const gridEl = viewportRef.current?.querySelector('.grid-preview') as HTMLElement | null;
+    if (!gridEl) return;
+    await exportToPdf(
+      gridEl,
+      bomItems,
+      { gridResult, spacerConfig, unitSystem, layoutName: layoutMeta.name },
+      () => setExportPdfError('PDF export failed. Please try again.'),
+    );
+  }, [bomItems, gridResult, spacerConfig, unitSystem, layoutMeta.name, setExportPdfError]);
+
+  const handleRemoveImage = (id: string) => {
+    removeRefImagePlacement(id);
+    if (selectedImageId === id) setSelectedImageId(null);
+  };
+
+  const handleRebindImage = useCallback((id: string) => {
+    dialogDispatch({ type: 'OPEN_REBIND', targetId: id });
+  }, [dialogDispatch]);
+
+  const toggleImageViewMode = useCallback(() => {
+    setImageViewMode(prev => {
+      const next = prev === 'ortho' ? 'perspective' : 'ortho';
+      localStorage.setItem('gridfinity-image-view-mode', next);
+      return next;
+    });
+  }, []);
+
+  // Keyboard shortcuts
+  const keyDownHandlerRef = useRef<((event: KeyboardEvent) => void) | undefined>(undefined);
+
+  keyDownHandlerRef.current = (event: KeyboardEvent) => {
+    const activeElement = document.activeElement;
+    const isTyping = activeElement?.tagName === 'INPUT' ||
+                     activeElement?.tagName === 'TEXTAREA' ||
+                     activeElement?.tagName === 'SELECT';
+    if (isTyping) return;
+
+    if ((event.key === 'Delete' || event.key === 'Backspace')) {
+      if (selectedImageId) {
+        event.preventDefault();
+        removeRefImagePlacement(selectedImageId);
+        setSelectedImageId(null);
+        return;
+      }
+      if (selectedItemIds.size > 0) { event.preventDefault(); deleteSelected(); return; }
+    }
+
+    if (event.key === 'r' || event.key === 'R') {
+      if (selectedImageId) {
+        event.preventDefault();
+        updateRefImageRotation(selectedImageId, event.shiftKey ? 'ccw' : 'cw');
+        return;
+      }
+      if (selectedItemIds.size > 0) {
+        event.preventDefault();
+        rotateSelected(event.shiftKey ? 'ccw' : 'cw');
+        return;
+      }
+    }
+
+    if ((event.key === 'd' || event.key === 'D') && (event.ctrlKey || event.metaKey)) {
+      event.preventDefault(); duplicateItem(); return;
+    }
+    if ((event.key === 'c' || event.key === 'C') && (event.ctrlKey || event.metaKey)) {
+      event.preventDefault(); copyItems(); return;
+    }
+    if ((event.key === 'v' || event.key === 'V') && (event.ctrlKey || event.metaKey)) {
+      event.preventDefault(); pasteItems(); return;
+    }
+    if ((event.key === 'v' || event.key === 'V') && !event.ctrlKey && !event.metaKey) {
+      event.preventDefault(); toggleImageViewMode(); return;
+    }
+    if (event.key === 'Escape') { deselectAll(); setSelectedImageId(null); return; }
+    if ((event.key === 'a' || event.key === 'A') && (event.ctrlKey || event.metaKey)) {
+      event.preventDefault(); selectAll(); return;
+    }
+    if ((event.key === 'l' || event.key === 'L') && selectedImageId) {
+      event.preventDefault(); toggleRefImageLock(selectedImageId); return;
+    }
+    if (event.key === '+' || event.key === '=') { event.preventDefault(); zoomIn(); return; }
+    if (event.key === '-' || event.key === '_') { event.preventDefault(); zoomOut(); return; }
+    if (event.key === '0' && (event.ctrlKey || event.metaKey)) {
+      event.preventDefault(); resetZoom(); return;
+    }
+    if (event.key === '?') {
+      event.preventDefault(); dialogDispatch({ type: 'TOGGLE', dialog: 'keyboard' }); return;
+    }
+    if (event.key === ' ') {
+      event.preventDefault();
+      isSpaceHeldRef.current = true;
+      if (viewportRef.current) viewportRef.current.style.cursor = 'grab';
+      return;
+    }
+  };
+
+  useEffect(() => {
+    const handler = (e: KeyboardEvent) => keyDownHandlerRef.current?.(e);
+    document.addEventListener('keydown', handler);
+
+    const handleKeyUp = (e: KeyboardEvent) => {
+      if (e.key === ' ') {
+        isSpaceHeldRef.current = false;
+        if (viewportRef.current) viewportRef.current.style.cursor = '';
+      }
+    };
+    document.addEventListener('keyup', handleKeyUp);
+
+    return () => {
+      document.removeEventListener('keydown', handler);
+      document.removeEventListener('keyup', handleKeyUp);
+    };
+  }, []);
+
+  const dimensionsContent = (
+    <>
+      <div className="unit-toggle-compact">
+        <button className={unitSystem === 'metric' ? 'active' : ''} onClick={() => handleUnitChange('metric')}>mm</button>
+        <button className={unitSystem === 'imperial' ? 'active' : ''} onClick={() => handleUnitChange('imperial')}>in</button>
+      </div>
+      {unitSystem === 'imperial' && (
+        <div className="format-toggle-compact">
+          <button className={imperialFormat === 'decimal' ? 'active' : ''} onClick={() => setImperialFormat('decimal')}>.00</button>
+          <button className={imperialFormat === 'fractional' ? 'active' : ''} onClick={() => setImperialFormat('fractional')}>½</button>
+        </div>
+      )}
+      <div className="dimension-inputs-row">
+        <DimensionInput label="Width" value={width} onChange={setWidth} unit={unitSystem} imperialFormat={imperialFormat} />
+        <span className="dimension-separator">x</span>
+        <DimensionInput label="Depth" value={depth} onChange={setDepth} unit={unitSystem} imperialFormat={imperialFormat} />
+      </div>
+      <GridSummary
+        gridX={gridResult.gridX} gridY={gridResult.gridY}
+        gapWidth={gridResult.gapWidth} gapDepth={gridResult.gapDepth}
+        unit={unitSystem} imperialFormat={imperialFormat}
+      />
+    </>
+  );
+
+  const spacerContent = (
+    <SpacerControls config={spacerConfig} onConfigChange={setSpacerConfig} />
+  );
+
+  return (
+    <>
+      <SidebarPanel
+        dimensionsContent={dimensionsContent}
+        spacerContent={spacerContent}
+        onClearCanvas={handleClearAll}
+        onReset={handleReset}
+        isReadOnly={isReadOnly}
+      />
+
+      <section className="preview">
+        <div className="preview-toolbar">
+          <div className="reference-image-toolbar">
+            {isAuthenticated && (
+              <button className="layout-toolbar-btn layout-load-btn" onClick={() => dialogDispatch({ type: 'OPEN', dialog: 'load' })} type="button">Load</button>
+            )}
+            {isAuthenticated && (
+              <button
+                className="layout-toolbar-btn layout-save-btn"
+                onClick={() => dialogDispatch({ type: 'OPEN', dialog: 'save' })}
+                type="button"
+                disabled={placedItems.length === 0 && refImagePlacements.length === 0}
+              >
+                Save
+              </button>
+            )}
+            {isAuthenticated && layoutMeta.status !== 'submitted' && layoutMeta.status !== 'delivered' && (
+              <button
+                className="layout-toolbar-btn layout-submit-btn"
+                onClick={handleSubmitClick}
+                type="button"
+                disabled={submitLayoutMutation.isPending}
+              >
+                {submitLayoutMutation.isPending ? 'Submitting...' : 'Submit'}
+              </button>
+            )}
+            {isAuthenticated && layoutMeta.status === 'delivered' && (
+              <button className="layout-toolbar-btn layout-submit-btn" disabled type="button" title="This layout has been fulfilled">
+                Submit
+              </button>
+            )}
+            {isAuthenticated && layoutMeta.id && layoutMeta.status === 'submitted' && (
+              <button className="layout-toolbar-btn layout-withdraw-btn" onClick={handleWithdrawLayout} type="button" disabled={withdrawLayoutMutation.isPending}>
+                {withdrawLayoutMutation.isPending ? 'Withdrawing...' : 'Withdraw'}
+              </button>
+            )}
+            {isAuthenticated && isReadOnly && (
+              <button className="layout-toolbar-btn layout-clone-btn" onClick={handleCloneCurrentLayout} type="button" disabled={cloneLayoutMutation.isPending}>
+                {cloneLayoutMutation.isPending ? 'Cloning...' : 'Clone'}
+              </button>
+            )}
+            <button
+              className="layout-toolbar-btn layout-export-btn"
+              onClick={handleExportPdf}
+              type="button"
+              disabled={placedItems.length === 0}
+              title="Export layout as PDF"
+            >
+              Export PDF
+            </button>
+            {exportPdfError && (
+              <span className="export-pdf-error" role="alert">{exportPdfError}</span>
+            )}
+            {!isReadOnly && (placedItems.length > 0 || refImagePlacements.length > 0) && (
+              <button className="clear-all-button" onClick={handleClearAll}>Clear All ({placedItems.length + refImagePlacements.length})</button>
+            )}
+            {isAdmin && (
+              <SubmissionsBadge count={submittedCountQuery.data?.submitted ?? 0} onClick={() => dialogDispatch({ type: 'OPEN', dialog: 'admin' })} />
+            )}
+          </div>
+          <ImageViewToggle mode={imageViewMode} onToggle={toggleImageViewMode} />
+          <ZoomControls zoom={transform.zoom} onZoomIn={zoomIn} onZoomOut={zoomOut} onResetZoom={resetZoom} onFitToScreen={handleFitToScreen} />
+        </div>
+        <GridViewport
+          viewportRef={viewportRef}
+          transform={transform}
+          handleWheel={handleWheel}
+          setZoomLevel={setZoomLevel}
+          pan={pan}
+          isSpaceHeldRef={isSpaceHeldRef}
+        >
+          <GridPreview
+            gridX={gridResult.gridX}
+            gridY={gridResult.gridY}
+            placedItems={placedItems}
+            selectedItemIds={selectedItemIds}
+            spacers={spacers}
+            imageViewMode={imageViewMode}
+            onDrop={handleCombinedDrop}
+            onSelectItem={(id, mods) => { selectItem(id, mods); if (id) setSelectedImageId(null); }}
+            getItemById={getItemById}
+            onDeleteItem={deleteItem}
+            onRotateItemCw={(id) => rotateItem(id, 'cw')}
+            onRotateItemCcw={(id) => rotateItem(id, 'ccw')}
+            onItemCustomizationChange={updateItemCustomization}
+            onItemCustomizationReset={(id) => updateItemCustomization(id, undefined)}
+            onDuplicateItem={duplicateItem}
+            getLibraryMeta={getLibraryMeta}
+            referenceImages={referenceImagesForGrid}
+            selectedImageId={selectedImageId}
+            onImagePositionChange={updateRefImagePosition}
+            onImageSelect={(id) => { setSelectedImageId(id); deselectAll(); }}
+            onImageScaleChange={updateRefImageScale}
+            onImageOpacityChange={updateRefImageOpacity}
+            onImageRemove={handleRemoveImage}
+            onImageToggleLock={toggleRefImageLock}
+            onImageRotateCw={(id) => updateRefImageRotation(id, 'cw')}
+            onImageRotateCcw={(id) => updateRefImageRotation(id, 'ccw')}
+            refImageMetadata={refImageMetadata}
+            onRefImageRebind={handleRebindImage}
+          />
+        </GridViewport>
+      </section>
+
+      <div className="library-resize-handle" onMouseDown={handleLibraryResizeStart} role="separator" aria-label="Resize library panel" />
+      <section className="library-panel" style={{ width: libraryWidth, minWidth: libraryWidth }}>
+        <div className="library-panel-header">
+          <div className="library-panel-header-icon">⊞</div>
+          <div className="library-panel-header-text">
+            <span className="library-panel-title">Component Library</span>
+            <span className="library-panel-subtitle">Drag to workspace</span>
+          </div>
+        </div>
+        <div className="library-panel-tabs">
+          <button
+            className={`library-cat-tab${libraryTab === 'items' && !libraryCategory ? ' active' : ''}`}
+            onClick={() => { setLibraryTab('items'); setLibraryCategory(null); }}
+            type="button"
+          >All</button>
+          {categories.map(cat => (
+            <button
+              key={cat.id}
+              className={`library-cat-tab${libraryTab === 'items' && libraryCategory === cat.id ? ' active' : ''}`}
+              onClick={() => { setLibraryTab('items'); setLibraryCategory(cat.id); }}
+              type="button"
+            >{cat.name}</button>
+          ))}
+          {isAuthenticated && (
+            <button
+              className={`library-cat-tab${libraryTab === 'images' ? ' active' : ''}`}
+              onClick={() => setLibraryTab('images')}
+              type="button"
+            >Images</button>
+          )}
+        </div>
+        <div className="library-panel-content">
+          {libraryTab === 'items' ? (
+            <>
+              <ItemLibrary
+                items={libraryItems}
+                categories={categories}
+                isLoading={isLibraryLoading || isLibrariesLoading}
+                error={libraryError || librariesError}
+                onRefreshLibrary={handleRefreshAll}
+                activeCategory={libraryCategory}
+              />
+              {isAuthenticated && <UserStlLibrarySection />}
+            </>
+          ) : isAuthenticated ? (
+            <RefImageLibrary />
+          ) : (
+            <div className="ref-image-auth-prompt">
+              <p>Sign in to upload and manage reference images.</p>
+            </div>
+          )}
+        </div>
+        <div className="library-panel-bom">
+          <BillOfMaterials items={bomItems} />
+        </div>
+        {selectedItemIds.size > 0 && (
+          <div className="library-panel-selection">
+            <ItemControls
+              onRotateCw={handleRotateSelectedCw}
+              onRotateCcw={handleRotateSelectedCcw}
+              onDelete={handleDeleteSelected}
+            />
+            {selectedItemIds.size === 1 && (() => {
+              const selectedId = selectedItemIds.values().next().value as string;
+              const selectedItem = placedItems.find(i => i.instanceId === selectedId);
+              if (!selectedItem) return null;
+              return (
+                <BinCustomizationPanel
+                  customization={selectedItem.customization}
+                  onChange={(c) => updateItemCustomization(selectedId, c)}
+                  onReset={() => updateItemCustomization(selectedId, undefined)}
+                  customizableFields={selectedLibraryMeta.customizableFields}
+                  customizationDefaults={selectedLibraryMeta.customizationDefaults}
+                />
+              );
+            })()}
+          </div>
+        )}
+      </section>
+
+      <LoadLayoutDialog
+        isOpen={dialogs.load}
+        onClose={() => dialogDispatch({ type: 'CLOSE', dialog: 'load' })}
+        onLoad={handleLoadLayout}
+        hasItems={placedItems.length > 0 || refImagePlacements.length > 0}
+      />
+    </>
+  );
+}
